requirements.txt
requests
pytest
PyYaml
tzlocal